util: Fix return type in macro TEMP_RETRY.
int was used as return type for the function read, write, recv, and send.
Now the correct type ssize_t is used to prevent overflows.

+++ src/util-io/io.h
@@ -40,14 +40,14 @@ typedef SSIZE_T ssize_t;
     dest = __ret; }
 #elif defined (__FreeBSD__)
 #define TEMP_RETRY(dest, exp) \
-{   int __ret; \
+{   ssize_t __ret; \
     do { \
         __ret = exp; \
     } while ((__ret == SOCKET_ERROR) && (errno == EINTR || errno == EAGAIN)); \
     dest =__ret; }
 #else
 #define TEMP_RETRY(dest, exp) \
-{   int __ret; \
+{   ssize_t __ret; \
     do { \
         __ret = exp; \
     } while (__ret == SOCKET_ERROR && errno == EINTR); \
